Move test_simplepre_ursula_worker_integration.py to actors/test_worker.py
